+++ src/index.ts
@@ -7,6 +7,7 @@ export { default as Input } from '@/package/components/input.vue';
 export { default as Textarea } from '@/package/components/textarea.vue';
 export { default as Select } from '@/package/components/select.vue';
 export { default as Autocomplete } from '@/package/components/autocomplete.vue';
+export { default as DatePicker } from '@/package/components/datepicker.vue';
 export { default as Checkbox } from '@/package/components/checkbox.vue';
 export { default as RadioGroup } from '@/package/components/radio-group.vue';
 export { default as RadioButton } from '@/package/components/radio-button.vue';